chatbot_api/pyproject.toml
[project]
name = "chatbot_api"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "langchain>=0.2.0",
  "langchain-community>=0.2.0",
  "langchain-google-genai>=2.0.0",
  "google-generativeai>=0.7.0",
  "neo4j>=5.20.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = ["ipython", "ruff", "black"]
